fix: update components and add test for island
