+++ src/modules/item/Item.tsx
@@ -60,7 +60,7 @@ import { useCurrentMemberContext } from '@/contexts/CurrentMemberContext';
 import { PLAYER } from '@/langs/constants';
 import { isHidden, paginationContentFilter } from '@/utils/item';
 
-import NavigationIsland from './NavigationIsland';
+import NavigationIsland from '../navigationIsland/NavigationIsland';
 import PinnedFolderItem from './PinnedFolderItem';
 
 const {
@@ -361,13 +361,7 @@ const ItemContent = ({ item }: ItemContentProps) => {
   }
 };
 
-const ItemContentWrapper = ({
-  item,
-  showIsland = false,
-}: {
-  item: DiscriminatedItem;
-  showIsland?: boolean;
-}) => {
+const ItemContentWrapper = ({ item }: { item: DiscriminatedItem }) => {
   const { data: itemTags } = useItemTags(item.id);
   const isItemHidden = isHidden(item, itemTags);
 
@@ -375,12 +369,7 @@ const ItemContentWrapper = ({
   if (isItemHidden) {
     return null;
   }
-  return (
-    <>
-      <ItemContent item={item} />
-      {showIsland && <NavigationIsland />}
-    </>
-  );
+  return <ItemContent item={item} />;
 };
 
 type Props = {
@@ -518,7 +507,12 @@ const Item = ({
   }
 
   // executed when item is a single child that is not a folder
-  return <ItemContentWrapper item={item} showIsland />;
+  return (
+    <>
+      <ItemContentWrapper item={item} />
+      <NavigationIsland />
+    </>
+  );
 };
 
 export default Item;
